refactor(db): split the 6210-line lib.rs Database impl into store/ submodules
The single Database inherent impl mixed lifecycle, CRUD, edge resolution,
graph queries, RAG, and LSP helpers in one file. Move each cluster (the
seams the banner comments already marked) into store/{lifecycle,crud,
resolution,queries,rag,lsp}.rs, each a separate impl Database block over
`use super::*`. Struct, schema consts, and free helpers stay in lib.rs.

Pure code movement: no API change, behavior byte-identical (137 tests pass
unchanged). lib.rs drops to 4058 lines. FILE_CHUNK_SIZE is now pub(crate)
so a sibling impl block can see it.
