+++ crates/cartog-db/src/lib.rs
@@ -749,2161 +749,9 @@ fn backup_before_destructive_migration(
     Ok(())
 }
 
-impl Database {
-    /// Open or create the database at the given path.
-    ///
-    /// `embedding_dim` sets the vector dimension for the sqlite-vec table.
-    /// If the stored dimension differs from the requested one, the vector index
-    /// is cleared and recreated (a re-index via `cartog rag index` is needed).
-    pub fn open(path: impl AsRef<std::path::Path>, embedding_dim: usize) -> DbResult<Self> {
-        register_sqlite_vec();
-        let db_path = path.as_ref();
-        // SQLite::open fails on a missing parent tree, so materialize `.cartog/`.
-        if let Some(parent) = db_path.parent() {
-            if !parent.as_os_str().is_empty() {
-                std::fs::create_dir_all(parent).map_err(|source| DbError::PrepareDir {
-                    path: parent.to_path_buf(),
-                    source,
-                })?;
-            }
-        }
-        let conn = Connection::open(db_path).map_err(|source| DbError::Open {
-            path: db_path.to_path_buf(),
-            source,
-        })?;
-        conn.execute_batch(&format!(
-            "PRAGMA journal_mode=WAL;
-             PRAGMA busy_timeout={BUSY_TIMEOUT_MS};
-             PRAGMA foreign_keys=ON;
-             PRAGMA synchronous=NORMAL;
-             PRAGMA cache_size=-65536;
-             PRAGMA temp_store=MEMORY;
-             PRAGMA mmap_size=268435456;"
-        ))
-        .map_err(DbError::Pragma)?;
-        conn.execute_batch(SCHEMA).map_err(DbError::Schema)?;
-        conn.execute_batch(RAG_SCHEMA).map_err(DbError::RagSchema)?;
-        backup_before_destructive_migration(&conn, db_path)?;
-        migrate(&conn);
-        // Partial index requires resolution_state (added in migration 3→4),
-        // so create it after migrate() rather than from SCHEMA.
-        conn.execute_batch(
-            "CREATE INDEX IF NOT EXISTS idx_edges_unresolved
-                 ON edges(file_path) WHERE resolution_state = 0",
-        )
-        .map_err(DbError::Schema)?;
-        handle_embedding_dimension(&conn, embedding_dim).map_err(DbError::EmbeddingDimension)?;
-        Ok(Self { conn, pinned: None })
-    }
-
-    /// Open an existing on-disk database in **read-write** mode without
-    /// running schema migrations or the embedding-fingerprint reconcile.
-    /// Used by the Phase 5 promoter: a secondary that detected its primary
-    /// died and validated the on-disk schema/fingerprint against its
-    /// pinned snapshot before claiming the slot. Re-running the migration
-    /// would re-trigger the SQLITE_BUSY race that the election was meant
-    /// to prevent.
-    ///
-    /// Verifies that `schema_version` still matches `SCHEMA_VERSION` to
-    /// guard against a race where another writer upgraded the schema
-    /// between the secondary's attach and this promotion. Returns
-    /// [`DbError::SchemaDrift`] in that case so the promoter aborts and
-    /// the MCP process exits cleanly.
-    pub fn open_existing_rw(path: impl AsRef<std::path::Path>) -> DbResult<Self> {
-        register_sqlite_vec();
-        let db_path = path.as_ref();
-        let conn = Connection::open(db_path).map_err(|source| DbError::Open {
-            path: db_path.to_path_buf(),
-            source,
-        })?;
-        conn.execute_batch(&format!(
-            "PRAGMA journal_mode=WAL;
-             PRAGMA busy_timeout={BUSY_TIMEOUT_MS};
-             PRAGMA foreign_keys=ON;
-             PRAGMA synchronous=NORMAL;
-             PRAGMA cache_size=-65536;
-             PRAGMA temp_store=MEMORY;
-             PRAGMA mmap_size=268435456;"
-        ))
-        .map_err(DbError::Pragma)?;
-
-        let stored_schema = read_schema_version(&conn)?;
-        if stored_schema != SCHEMA_VERSION {
-            return Err(DbError::SchemaDrift {
-                expected: SCHEMA_VERSION,
-                stored: stored_schema,
-            });
-        }
-
-        Ok(Self { conn, pinned: None })
-    }
-
-    /// Open an existing on-disk database in **read-only** mode for a
-    /// secondary cartog process (Phase 4 read-only attach). Skips schema
-    /// migrations and the embedding-fingerprint reconcile — the primary
-    /// writer owns those.
-    ///
-    /// Behaviour:
-    /// - Opens with `SQLITE_OPEN_READ_ONLY` so write attempts surface as
-    ///   `SQLITE_READONLY` errors at runtime (a defense-in-depth backup
-    ///   for the higher-level tool gating).
-    /// - Reads the `metadata` snapshot (schema version + embedding
-    ///   fingerprint) and stores it on the returned [`Database`] so the
-    ///   promoter (Phase 5) can compare against the on-disk values later.
-    /// - Returns [`DbError::SchemaDrift`] if the stored `schema_version`
-    ///   doesn't match this binary's expected version — the primary
-    ///   upgraded cartog underneath us and queries can't be trusted.
-    pub fn open_readonly(path: impl AsRef<std::path::Path>) -> DbResult<Self> {
-        use rusqlite::OpenFlags;
-        register_sqlite_vec();
-        let db_path = path.as_ref();
-        let conn = Connection::open_with_flags(
-            db_path,
-            OpenFlags::SQLITE_OPEN_READ_ONLY | OpenFlags::SQLITE_OPEN_NO_MUTEX,
-        )
-        .map_err(|source| DbError::Open {
-            path: db_path.to_path_buf(),
-            source,
-        })?;
-        // busy_timeout is still useful: a long read can stall against a
-        // writer mid-checkpoint. WAL keeps readers and writers from
-        // blocking otherwise, but the timeout makes the bound explicit.
-        conn.execute_batch(&format!("PRAGMA busy_timeout={BUSY_TIMEOUT_MS};"))
-            .map_err(DbError::Pragma)?;
-
-        let stored_schema = read_schema_version(&conn)?;
-        if stored_schema != SCHEMA_VERSION {
-            return Err(DbError::SchemaDrift {
-                expected: SCHEMA_VERSION,
-                stored: stored_schema,
-            });
-        }
-
-        let stored_provider: Option<String> = conn
-            .query_row(
-                "SELECT value FROM metadata WHERE key = ?1",
-                params![EMBED_PROVIDER_KEY],
-                |row| row.get(0),
-            )
-            .optional()
-            .map_err(DbError::Sqlite)?;
-        let stored_model: Option<String> = conn
-            .query_row(
-                "SELECT value FROM metadata WHERE key = ?1",
-                params![EMBED_MODEL_KEY],
-                |row| row.get(0),
-            )
-            .optional()
-            .map_err(DbError::Sqlite)?;
-        let stored_dim: Option<usize> = conn
-            .query_row(
-                "SELECT CAST(value AS INTEGER) FROM metadata WHERE key = 'embedding_dimension'",
-                [],
-                |row| row.get::<_, i64>(0).map(|v| v as usize),
-            )
-            .optional()
-            .map_err(DbError::Sqlite)?;
-        // Embedding fingerprint is recorded together (Phase 6b backfill).
-        // If any field is missing the fingerprint is "unknown" — readers
-        // can still serve graph queries, just can't validate against a
-        // promoter swap later.
-        let embedding = match (stored_provider, stored_model, stored_dim) {
-            (Some(provider), Some(model), Some(dimension)) => Some(EmbeddingFingerprint {
-                provider,
-                model,
-                dimension,
-            }),
-            _ => None,
-        };
-
-        Ok(Self {
-            conn,
-            pinned: Some(PinnedAttach {
-                schema_version: stored_schema,
-                embedding,
-            }),
-        })
-    }
-
-    /// Open an in-memory database (for tests and benchmarks).
-    #[doc(hidden)]
-    pub fn open_memory() -> DbResult<Self> {
-        register_sqlite_vec();
-        let conn = Connection::open_in_memory()?;
-        conn.execute_batch("PRAGMA foreign_keys=ON;")
-            .map_err(DbError::Pragma)?;
-        conn.execute_batch(SCHEMA).map_err(DbError::Schema)?;
-        conn.execute_batch(RAG_SCHEMA).map_err(DbError::RagSchema)?;
-        conn.execute_batch(&rag_vec_schema(DEFAULT_EMBEDDING_DIM))
-            .map_err(DbError::RagSchema)?;
-        migrate(&conn);
-        conn.execute_batch(
-            "CREATE INDEX IF NOT EXISTS idx_edges_unresolved
-                 ON edges(file_path) WHERE resolution_state = 0",
-        )
-        .map_err(DbError::Schema)?;
-        Ok(Self { conn, pinned: None })
-    }
-
-    /// True when this `Database` was opened via [`Self::open_readonly`].
-    /// MCP tool gating (Phase 4) consults this to refuse the 2 write tools.
-    pub fn is_read_only(&self) -> bool {
-        self.pinned.is_some()
-    }
-
-    /// Snapshot captured at attach time when [`Self::open_readonly`] was
-    /// used. `None` for read-write opens.
-    pub fn pinned_attach(&self) -> Option<&PinnedAttach> {
-        self.pinned.as_ref()
-    }
-
-    /// Cap the number of pages this DB connection can hold.
-    ///
-    /// Intended for tests that need to force a `SQLITE_FULL` error on a
-    /// subsequent write (for example, to verify that a transaction rolls back
-    /// cleanly). Production code should never call this.
-    #[doc(hidden)]
-    pub fn set_max_page_count_for_tests(&self, pages: u32) -> Result<()> {
-        self.conn
-            .execute_batch(&format!("PRAGMA max_page_count = {pages}"))?;
-        Ok(())
-    }
-
-    /// Open a single SQLite transaction that the caller is expected to wrap
-    /// around a multi-step indexing pipeline.
-    ///
-    /// Drop without `commit()` rolls back, so a panic mid-pipeline leaves the
-    /// DB in its prior state.
-    ///
-    /// # Calling conventions inside the transaction
-    ///
-    /// Helpers fall into two categories:
-    ///
-    /// 1. **Batched writers must use the `_in_tx` variant.** Their non-`_in_tx`
-    ///    wrapper issues its own `BEGIN` and would error out at runtime
-    ///    (`cannot start a transaction within a transaction`). Examples:
-    ///    [`Self::insert_symbols_in_tx`], [`Self::delete_symbols_in_tx`],
-    ///    [`Self::insert_edges_in_tx`], [`Self::insert_symbol_contents_in_tx`],
-    ///    [`Self::clear_file_data_in_tx`], [`Self::remove_file_in_tx`],
-    ///    [`Self::resolve_edges_in_tx`], [`Self::resolve_edges_scoped_in_tx`].
-    ///
-    /// 2. **Single-statement helpers can be called directly.** They issue one
-    ///    `self.conn.execute(...)` and participate transparently in the active
-    ///    transaction. Examples used by `cartog-indexer`'s Phase 3 today:
-    ///    [`Self::upsert_file`], [`Self::clear_edges_for_file`],
-    ///    [`Self::set_metadata`], [`Self::compute_in_degrees`],
-    ///    [`Self::compute_in_degrees_scoped`], [`Self::invalidate_edges_targeting`].
-    ///    These are tagged with `// tx-safe: single statement` so the contract
-    ///    survives drive-by edits.
-    ///
-    /// # Why `unchecked_transaction` rather than [`rusqlite::Connection::transaction`]
-    ///
-    /// `transaction()` requires `&mut Connection`, which would force every
-    /// caller of `Database` to hold a mutable borrow for the entire pipeline.
-    /// `unchecked_transaction()` works through `&Connection` and produces an
-    /// equivalent [`rusqlite::Transaction`] with the same `DropBehavior::Rollback`
-    /// default — only borrow-check ergonomics differ.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if SQLite cannot begin a transaction — typically
-    /// because another transaction is already active on this connection.
-    pub fn begin_indexing_tx(&self) -> Result<rusqlite::Transaction<'_>> {
-        Ok(self.conn.unchecked_transaction()?)
-    }
-
-    // ── Metadata ──
-
-    /// Retrieve a metadata value by key.
-    pub fn get_metadata(&self, key: &str) -> Result<Option<String>> {
-        self.conn
-            .query_row(
-                "SELECT value FROM metadata WHERE key = ?1",
-                params![key],
-                |row| row.get(0),
-            )
-            .optional()
-            .context("Failed to query metadata")
-    }
-
-    /// Store a metadata key-value pair (upserts on conflict).
-    ///
-    /// tx-safe: single statement — see [`Self::begin_indexing_tx`].
-    pub fn set_metadata(&self, key: &str, value: &str) -> Result<()> {
-        self.conn.execute(
-            "INSERT OR REPLACE INTO metadata (key, value) VALUES (?1, ?2)",
-            params![key, value],
-        )?;
-        Ok(())
-    }
-
-    /// Reconcile the stored embedding fingerprint with the one currently in
-    /// use. Call right after `Database::open` from any code path that owns
-    /// an `EmbeddingProvider` (indexer, watcher, MCP serve, `rag index`).
-    ///
-    /// Behavior:
-    /// - All three fields (provider, model, dimension) match stored → no-op,
-    ///   zero writes.
-    /// - Any field differs → drop `symbol_vec`, clear `symbol_embedding_map`,
-    ///   recreate the vector table at the new dimension, update all three
-    ///   metadata keys. The user must run `cartog rag index` to repopulate.
-    /// - DB has dimension but no provider/model (older cartog versions) →
-    ///   backfill provider+model without wiping. The stored vectors stay
-    ///   valid against whatever stack produced them; we just record the
-    ///   identity going forward.
-    ///
-    /// Writes use [`retry_busy`] so a concurrent writer on the same DB does
-    /// not crash this caller with `SQLITE_BUSY`.
-    pub fn reconcile_embedding_fingerprint(&self, fp: &EmbeddingFingerprint) -> Result<()> {
-        let stored_provider: Option<String> = self.get_metadata(EMBED_PROVIDER_KEY)?;
-        let stored_model: Option<String> = self.get_metadata(EMBED_MODEL_KEY)?;
-        let stored_dim: Option<usize> = self
-            .conn
-            .query_row(
-                "SELECT CAST(value AS INTEGER) FROM metadata WHERE key = 'embedding_dimension'",
-                [],
-                |row| row.get::<_, i64>(0).map(|v| v as usize),
-            )
-            .optional()
-            .context("Failed to query embedding dimension")?;
-
-        // Full match AND `symbol_vec` actually exists on disk: zero writes.
-        // The dim+table pair is the real invariant; checking metadata
-        // alone misses the case where a previous open crashed mid-
-        // migration and left the DB without `symbol_vec` while metadata
-        // still claims a fingerprint.
-        if stored_provider.as_deref() == Some(fp.provider.as_str())
-            && stored_model.as_deref() == Some(fp.model.as_str())
-            && stored_dim == Some(fp.dimension)
-            && symbol_vec_exists(&self.conn)?
-        {
-            return Ok(());
-        }
-
-        // Backwards-compat: stored has dim from an older cartog (no provider/model
-        // recorded yet). Treat first-time-with-provider as a backfill, not an
-        // invalidation. Embeddings produced by the previous run are still valid
-        // against whatever stack the user had configured then.
-        let dim_matches = stored_dim == Some(fp.dimension);
-        let is_backfill = dim_matches && stored_provider.is_none() && stored_model.is_none();
-
-        if !is_backfill {
-            tracing::warn!(
-                old_provider = ?stored_provider,
-                old_model = ?stored_model,
-                old_dim = ?stored_dim,
-                new_provider = %fp.provider,
-                new_model = %fp.model,
-                new_dim = fp.dimension,
-                "Embedding fingerprint changed — clearing vector index. Run `cartog rag index` to re-embed."
-            );
-        }
-
-        // Wrap the whole sequence in a transaction so a mid-sequence
-        // failure (e.g. busy-retry exhausted on the third metadata write)
-        // rolls back atomically. Otherwise the next open could see
-        // partial state — e.g. provider/model match but dimension stale,
-        // or symbol_vec dropped but metadata still pointing at the old
-        // dim — and either skip migration or silently re-wipe.
-        let schema = rag_vec_schema(fp.dimension);
-        let do_wipe = !is_backfill;
-        retry_busy(|| {
-            let tx = self.conn.unchecked_transaction()?;
-            if do_wipe {
-                tx.execute("DROP TABLE IF EXISTS symbol_vec", [])?;
-                tx.execute("DELETE FROM symbol_embedding_map", [])?;
-            }
-            tx.execute_batch(&schema)?;
-            tx.execute(
-                "INSERT OR REPLACE INTO metadata (key, value) VALUES (?1, ?2)",
-                params![EMBED_PROVIDER_KEY, fp.provider],
-            )?;
-            tx.execute(
-                "INSERT OR REPLACE INTO metadata (key, value) VALUES (?1, ?2)",
-                params![EMBED_MODEL_KEY, fp.model],
-            )?;
-            #[cfg(test)]
-            if RECONCILE_FAIL_AFTER_MODEL
-                .with(|b| b.swap(false, std::sync::atomic::Ordering::SeqCst))
-            {
-                return Err(rusqlite::Error::SqliteFailure(
-                    rusqlite::ffi::Error::new(rusqlite::ffi::SQLITE_FULL),
-                    Some("injected mid-sequence failure".into()),
-                ));
-            }
-            tx.execute(
-                "INSERT OR REPLACE INTO metadata (key, value) VALUES ('embedding_dimension', ?1)",
-                params![fp.dimension.to_string()],
-            )?;
-            tx.commit()
-        })
-        .map_err(|e| anyhow::anyhow!("failed to reconcile embedding fingerprint: {e}"))?;
-
-        Ok(())
-    }
-
-    // ── Files ──
-
-    /// Insert or update file metadata.
-    ///
-    /// tx-safe: single statement — see [`Self::begin_indexing_tx`].
-    pub fn upsert_file(&self, file: &FileInfo) -> Result<()> {
-        self.conn.execute(
-            "INSERT OR REPLACE INTO files (path, last_modified, hash, language, num_symbols)
-             VALUES (?1, ?2, ?3, ?4, ?5)",
-            params![
-                file.path,
-                file.last_modified,
-                file.hash,
-                file.language,
-                file.num_symbols,
-            ],
-        )?;
-        Ok(())
-    }
-
-    /// Look up stored metadata for a file.
-    pub fn get_file(&self, path: &str) -> Result<Option<FileInfo>> {
-        self.conn
-            .query_row(
-                "SELECT path, last_modified, hash, language, num_symbols FROM files WHERE path = ?1",
-                params![path],
-                |row| {
-                    Ok(FileInfo {
-                        path: row.get(0)?,
-                        last_modified: row.get(1)?,
-                        hash: row.get(2)?,
-                        language: row.get(3)?,
-                        num_symbols: row.get(4)?,
-                    })
-                },
-            )
-            .optional()
-            .context("Failed to query file")
-    }
-
-    /// Remove edges only for a file (used by Merkle diff which updates symbols surgically).
-    ///
-    /// tx-safe: single statement — see [`Self::begin_indexing_tx`].
-    pub fn clear_edges_for_file(&self, path: &str) -> Result<()> {
-        self.conn
-            .execute("DELETE FROM edges WHERE file_path = ?1", params![path])?;
-        Ok(())
-    }
-
-    /// Remove all symbols, edges, and RAG data for a file (before re-indexing it).
-    pub fn clear_file_data(&self, path: &str) -> Result<()> {
-        let tx = self.conn.unchecked_transaction()?;
-        self.clear_file_data_in_tx(path)?;
-        tx.commit()?;
-        Ok(())
-    }
-
-    /// Like [`Self::clear_file_data`] but assumes the caller already holds an
-    /// open transaction. Used by `cartog-indexer` to wrap the entire Phase 3
-    /// pipeline atomically.
-    pub fn clear_file_data_in_tx(&self, path: &str) -> Result<()> {
-        self.clear_rag_data_for_file(path)?;
-        self.conn
-            .execute("DELETE FROM edges WHERE file_path = ?1", params![path])?;
-        self.conn
-            .execute("DELETE FROM symbols WHERE file_path = ?1", params![path])?;
-        Ok(())
-    }
-
-    /// Remove a file and all its symbols and edges from the index.
-    pub fn remove_file(&self, path: &str) -> Result<()> {
-        let tx = self.conn.unchecked_transaction()?;
-        self.remove_file_in_tx(path)?;
-        tx.commit()?;
-        Ok(())
-    }
-
-    /// Like [`Self::remove_file`] but assumes the caller already holds an
-    /// open transaction.
-    pub fn remove_file_in_tx(&self, path: &str) -> Result<()> {
-        self.clear_file_data_in_tx(path)?;
-        self.conn
-            .execute("DELETE FROM files WHERE path = ?1", params![path])?;
-        Ok(())
-    }
-
-    // ── Symbols ──
-
-    /// Insert or replace a single symbol.
-    #[cfg_attr(not(test), allow(dead_code))]
-    pub fn insert_symbol(&self, sym: &Symbol) -> Result<()> {
-        self.conn
-            .prepare_cached(SQL_INSERT_SYMBOL)?
-            .execute(params![
-                sym.id,
-                sym.name,
-                sym.kind.as_str(),
-                sym.file_path,
-                sym.start_line,
-                sym.end_line,
-                sym.start_byte,
-                sym.end_byte,
-                sym.parent_id,
-                sym.signature,
-                sym.visibility.as_str(),
-                sym.is_async,
-                sym.docstring,
-                sym.content_hash,
-                sym.subtree_hash,
-            ])?;
-        Ok(())
-    }
-
-    /// Insert or replace multiple symbols in a single transaction.
-    pub fn insert_symbols(&self, symbols: &[Symbol]) -> Result<()> {
-        let tx = self.conn.unchecked_transaction()?;
-        self.insert_symbols_in_tx(symbols)?;
-        tx.commit()?;
-        Ok(())
-    }
-
-    /// Like [`Self::insert_symbols`] but assumes the caller already holds an
-    /// open transaction.
-    pub fn insert_symbols_in_tx(&self, symbols: &[Symbol]) -> Result<()> {
-        let mut stmt = self.conn.prepare_cached(SQL_INSERT_SYMBOL)?;
-        for sym in symbols {
-            stmt.execute(params![
-                sym.id,
-                sym.name,
-                sym.kind.as_str(),
-                sym.file_path,
-                sym.start_line,
-                sym.end_line,
-                sym.start_byte,
-                sym.end_byte,
-                sym.parent_id,
-                sym.signature,
-                sym.visibility.as_str(),
-                sym.is_async,
-                sym.docstring,
-                sym.content_hash,
-                sym.subtree_hash,
-            ])?;
-        }
-        Ok(())
-    }
-
-    /// Get stored symbol hashes for a file (for Merkle diff).
-    /// Returns `(id, content_hash, subtree_hash)` tuples.
-    #[allow(clippy::type_complexity)]
-    pub fn get_symbol_hashes_for_file(
-        &self,
-        file_path: &str,
-    ) -> Result<Vec<(String, Option<String>, Option<String>)>> {
-        let mut stmt = self
-            .conn
-            .prepare("SELECT id, content_hash, subtree_hash FROM symbols WHERE file_path = ?1")?;
-        let rows = stmt
-            .query_map(params![file_path], |row| {
-                Ok((row.get(0)?, row.get(1)?, row.get(2)?))
-            })?
-            .collect::<std::result::Result<Vec<_>, _>>()?;
-        Ok(rows)
-    }
-
-    /// Update only the position fields of a symbol (for moved-but-unchanged symbols).
-    pub fn update_symbol_position(
-        &self,
-        id: &str,
-        start_line: u32,
-        end_line: u32,
-        start_byte: u32,
-        end_byte: u32,
-    ) -> Result<()> {
-        self.conn.execute(
-            "UPDATE symbols SET start_line = ?2, end_line = ?3,
-                    start_byte = ?4, end_byte = ?5 WHERE id = ?1",
-            params![id, start_line, end_line, start_byte, end_byte],
-        )?;
-        Ok(())
-    }
-
-    /// Delete multiple symbols and cascade (edges, content, embeddings) in a
-    /// single transaction.
-    pub fn delete_symbols(&self, ids: &[String]) -> Result<()> {
-        if ids.is_empty() {
-            return Ok(());
-        }
-        let tx = self.conn.unchecked_transaction()?;
-        self.delete_symbols_in_tx(ids)?;
-        tx.commit()?;
-        Ok(())
-    }
-
-    /// Like [`Self::delete_symbols`] but assumes the caller already holds an
-    /// open transaction.
-    pub fn delete_symbols_in_tx(&self, ids: &[String]) -> Result<()> {
-        if ids.is_empty() {
-            return Ok(());
-        }
-        let mut del_out = self
-            .conn
-            .prepare_cached("DELETE FROM edges WHERE source_id = ?1")?;
-        // Reset state alongside target_id so the orphaned edge re-enters
-        // unresolved_edges() instead of becoming a (NULL, state=1) zombie.
-        let mut null_in = self.conn.prepare_cached(
-            "UPDATE edges SET target_id = NULL, resolution_state = 0 WHERE target_id = ?1",
-        )?;
-        let mut del_vec = self.conn.prepare_cached(
-            "DELETE FROM symbol_vec WHERE rowid IN \
-             (SELECT id FROM symbol_embedding_map WHERE symbol_id = ?1)",
-        )?;
-        let mut del_map = self
-            .conn
-            .prepare_cached("DELETE FROM symbol_embedding_map WHERE symbol_id = ?1")?;
-        let mut del_content = self
-            .conn
-            .prepare_cached("DELETE FROM symbol_content WHERE symbol_id = ?1")?;
-        let mut del_sym = self
-            .conn
-            .prepare_cached("DELETE FROM symbols WHERE id = ?1")?;
-        for id in ids {
-            del_out.execute(params![id])?;
-            null_in.execute(params![id])?;
-            del_vec.execute(params![id])?;
-            del_map.execute(params![id])?;
-            del_content.execute(params![id])?;
-            del_sym.execute(params![id])?;
-        }
-        Ok(())
-    }
-
-    /// Delete a single symbol and cascade to edges, content, and embeddings.
-    pub fn delete_symbol(&self, id: &str) -> Result<()> {
-        let tx = self.conn.unchecked_transaction()?;
-        self.conn
-            .execute("DELETE FROM edges WHERE source_id = ?1", params![id])?;
-        self.conn.execute(
-            "UPDATE edges SET target_id = NULL, resolution_state = 0 WHERE target_id = ?1",
-            params![id],
-        )?;
-        let _ = self.conn.execute(
-            "DELETE FROM symbol_vec WHERE rowid IN \
-             (SELECT id FROM symbol_embedding_map WHERE symbol_id = ?1)",
-            params![id],
-        );
-        let _ = self.conn.execute(
-            "DELETE FROM symbol_embedding_map WHERE symbol_id = ?1",
-            params![id],
-        );
-        let _ = self.conn.execute(
-            "DELETE FROM symbol_content WHERE symbol_id = ?1",
-            params![id],
-        );
-        self.conn
-            .execute("DELETE FROM symbols WHERE id = ?1", params![id])?;
-        tx.commit()?;
-        Ok(())
-    }
-
-    // ── Edges ──
-
-    /// Insert a single edge.
-    #[cfg_attr(not(test), allow(dead_code))]
-    pub fn insert_edge(&self, edge: &Edge) -> Result<()> {
-        self.conn.prepare_cached(SQL_INSERT_EDGE)?.execute(params![
-            edge.source_id,
-            edge.target_name,
-            edge.target_id,
-            edge.kind.as_str(),
-            edge.file_path,
-            edge.line,
-        ])?;
-        Ok(())
-    }
-
-    /// Insert multiple edges in a single transaction.
-    pub fn insert_edges(&self, edges: &[Edge]) -> Result<()> {
-        let tx = self.conn.unchecked_transaction()?;
-        self.insert_edges_in_tx(edges)?;
-        tx.commit()?;
-        Ok(())
-    }
-
-    /// Like [`Self::insert_edges`] but assumes the caller already holds an
-    /// open transaction.
-    pub fn insert_edges_in_tx(&self, edges: &[Edge]) -> Result<()> {
-        let mut stmt = self.conn.prepare_cached(SQL_INSERT_EDGE)?;
-        for edge in edges {
-            stmt.execute(params![
-                edge.source_id,
-                edge.target_name,
-                edge.target_id,
-                edge.kind.as_str(),
-                edge.file_path,
-                edge.line,
-            ])?;
-        }
-        Ok(())
-    }
-
-    // ── Edge Resolution ──
-
-    /// Resolve target_name → target_id for all unresolved edges.
-    ///
-    /// Runs two passes so that import edges resolved in pass 1 enable
-    /// import-path resolution (tier 2) for non-import edges in pass 2.
-    ///
-    /// 6-tier priority resolution (per pass):
-    /// 1. Same file — symbol with matching name in the same file
-    /// 2. Import-path — follow resolved imports to find the target in the imported file
-    /// 3. Same directory — symbol in a file in the same directory
-    /// 4. Parent scope preference — when multiple global matches, prefer same parent scope
-    /// 5. Unique project-wide match — exactly one symbol with that name globally
-    /// 6. Class over constructor — when exactly 2 matches and one is a class, prefer class
-    pub fn resolve_edges(&self) -> Result<u32> {
-        let tx = self.conn.unchecked_transaction()?;
-        let total = self.resolve_edges_in_tx()?;
-        tx.commit()?;
-        Ok(total)
-    }
-
-    /// Like [`Self::resolve_edges`] but assumes the caller already holds an
-    /// open transaction.
-    pub fn resolve_edges_in_tx(&self) -> Result<u32> {
-        let mut total_resolved = 0u32;
-        for _pass in 0..2 {
-            let resolved = self.resolve_edges_pass()?;
-            if resolved == 0 {
-                break;
-            }
-            total_resolved += resolved;
-        }
-        Ok(total_resolved)
-    }
-
-    fn resolve_edges_pass(&self) -> Result<u32> {
-        // Skip state=2 (unresolvable) edges: the heuristic can't resolve what
-        // LSP definitively gave up on, and re-walking them every dirty reindex
-        // burns measurable time on large repos. They re-enter via
-        // `reset_unresolvable_for_names` when a matching symbol appears.
-        let mut unresolved_stmt = self.conn.prepare(
-            "SELECT e.id, e.target_name, e.file_path, e.source_id
-             FROM edges e WHERE e.target_id IS NULL AND e.resolution_state = 0",
-        )?;
-
-        let unresolved: Vec<(i64, String, String, String)> = unresolved_stmt
-            .query_map([], |row| {
-                Ok((row.get(0)?, row.get(1)?, row.get(2)?, row.get(3)?))
-            })?
-            .collect::<std::result::Result<Vec<_>, _>>()?;
-
-        self.resolve_edge_batch(&unresolved)
-    }
-
-    /// 6-tier heuristic resolution for a batch of unresolved edges.
-    ///
-    /// Caller is responsible for transaction wrapping — see the public
-    /// [`Self::resolve_edges`] / [`Self::resolve_edges_scoped`] helpers.
-    fn resolve_edge_batch(&self, unresolved: &[(i64, String, String, String)]) -> Result<u32> {
-        let mut resolved = 0u32;
-
-        let mut same_file_stmt = self
-            .conn
-            .prepare("SELECT id FROM symbols WHERE name = ?1 AND file_path = ?2 LIMIT 1")?;
-
-        let mut import_resolve_stmt = self.conn.prepare(
-            "SELECT s.id FROM symbols s
-             INNER JOIN edges ie ON ie.kind = 'imports' AND ie.target_name = ?1
-                 AND ie.target_id IS NOT NULL
-             INNER JOIN symbols is2 ON is2.id = ie.source_id AND is2.file_path = ?2
-             INNER JOIN symbols resolved ON resolved.id = ie.target_id
-             WHERE s.name = ?1 AND s.kind != 'import'
-                 AND s.file_path = resolved.file_path
-             LIMIT 1",
-        )?;
-
-        let mut same_dir_stmt = self
-            .conn
-            .prepare("SELECT id FROM symbols WHERE name = ?1 AND file_path LIKE ?2 LIMIT 1")?;
-
-        let mut parent_scope_stmt = self.conn.prepare(
-            "SELECT s.id FROM symbols s
-             INNER JOIN symbols source ON source.id = ?2
-             WHERE s.name = ?1 AND s.parent_id = source.parent_id AND s.id != source.id
-             LIMIT 1",
-        )?;
-
-        let mut anywhere_stmt = self
-            .conn
-            .prepare("SELECT id, kind FROM symbols WHERE name = ?1 AND kind != 'import' LIMIT 3")?;
-
-        // Heuristic resolve must flip state=1 alongside target_id; otherwise
-        // unresolved_edges() (state=0 filter) would still surface the edge to
-        // the next LSP pass, wasting a server roundtrip on already-known answers.
-        let mut update_stmt = self
-            .conn
-            .prepare("UPDATE edges SET target_id = ?1, resolution_state = 1 WHERE id = ?2")?;
-
-        for (edge_id, target_name, edge_file, source_id) in unresolved {
-            let simple_name = target_name.rsplit('.').next().unwrap_or(target_name);
-
-            // 1) Same file
-            let target_id: Option<String> = same_file_stmt
-                .query_row(params![simple_name, edge_file], |row| row.get(0))
-                .optional()?;
-
-            if let Some(tid) = target_id {
-                update_stmt.execute(params![tid, edge_id])?;
-                resolved += 1;
-                continue;
-            }
-
-            // 2) Import-path resolution
-            let target_id: Option<String> = import_resolve_stmt
-                .query_row(params![simple_name, edge_file], |row| row.get(0))
-                .optional()?;
-
-            if let Some(tid) = target_id {
-                update_stmt.execute(params![tid, edge_id])?;
-                resolved += 1;
-                continue;
-            }
-
-            // 3) Same directory
-            let dir = edge_file
-                .rsplit_once('/')
-                .map(|(d, _)| format!("{d}/%"))
-                .unwrap_or_default();
-
-            if !dir.is_empty() {
-                let target_id: Option<String> = same_dir_stmt
-                    .query_row(params![simple_name, dir], |row| row.get(0))
-                    .optional()?;
-
-                if let Some(tid) = target_id {
-                    update_stmt.execute(params![tid, edge_id])?;
-                    resolved += 1;
-                    continue;
-                }
-            }
-
-            // 4) Parent scope preference
-            let target_id: Option<String> = parent_scope_stmt
-                .query_row(params![simple_name, source_id], |row| row.get(0))
-                .optional()?;
-
-            if let Some(tid) = target_id {
-                update_stmt.execute(params![tid, edge_id])?;
-                resolved += 1;
-                continue;
-            }
-
-            // 5+6) Project-wide: unique match, or class-over-constructor disambiguation
-            let mut rows = anywhere_stmt.query(params![simple_name])?;
-            let mut matches: Vec<(String, String)> = Vec::new();
-            while let Some(row) = rows.next()? {
-                matches.push((row.get(0)?, row.get(1)?));
-                if matches.len() == 3 {
-                    break;
-                }
-            }
-            drop(rows);
-
-            let resolved_id = match matches.len() {
-                1 => Some(&matches[0].0),
-                2 => disambiguate_two(&matches[0], &matches[1]),
-                _ => None,
-            };
-
-            if let Some(tid) = resolved_id {
-                update_stmt.execute(params![tid, edge_id])?;
-                resolved += 1;
-            }
-        }
-
-        Ok(resolved)
-    }
-
-    /// Compute and store in-degree centrality for all symbols.
-    ///
-    /// In-degree = number of resolved incoming edges (calls, imports, inherits, etc.).
-    /// Higher in-degree means the symbol is referenced more across the codebase.
-    /// Resets all in-degree values to 0 first, then batch-updates from the edges table.
-    ///
-    /// tx-safe: two unconditional statements participate in any active outer
-    /// transaction — see [`Self::begin_indexing_tx`].
-    pub fn compute_in_degrees(&self) -> Result<u32> {
-        self.conn.execute("UPDATE symbols SET in_degree = 0", [])?;
-
-        // CTE computes counts once; the UPDATE applies them.
-        // Avoids a correlated subquery per symbol (O(n*m) → O(n+m)).
-        let updated = self.conn.execute(
-            "WITH counts AS (
-                SELECT target_id, COUNT(*) AS cnt
-                FROM edges WHERE target_id IS NOT NULL
-                GROUP BY target_id
-            )
-            UPDATE symbols SET in_degree = (
-                SELECT cnt FROM counts WHERE counts.target_id = symbols.id
-            )
-            WHERE id IN (SELECT target_id FROM counts)",
-            [],
-        )?;
-
-        Ok(updated as u32)
-    }
-
-    // ── Scoped resolution (incremental indexing) ──
-
-    /// Invalidate resolved edges that point into any of the dirty files.
-    ///
-    /// When a file is re-indexed, its symbols may have been renamed/removed.
-    /// Edges from *unchanged* files that previously resolved to those symbols
-    /// must be cleared so they can be re-resolved against the new symbol set.
-    ///
-    /// tx-safe: single statement — see [`Self::begin_indexing_tx`].
-    pub fn invalidate_edges_targeting(
-        &self,
-        dirty_files: &std::collections::HashSet<String>,
-    ) -> Result<u32> {
-        if dirty_files.is_empty() {
-            return Ok(0);
-        }
-        // After file re-indexing, edges from unchanged files may point to
-        // symbol IDs that no longer exist (removed or renamed symbols).
-        // Set these dangling references to NULL AND reset resolution_state so
-        // the heuristic + LSP passes get another shot. Without the state reset
-        // the edge would stay at state=1 (resolved) but with target_id NULL —
-        // permanently invisible to `unresolved_edges()`.
-        let n = self.conn.execute(
-            "UPDATE edges SET target_id = NULL, resolution_state = 0
-             WHERE target_id IS NOT NULL
-               AND NOT EXISTS (SELECT 1 FROM symbols WHERE symbols.id = edges.target_id)",
-            [],
-        )?;
-        Ok(n as u32)
-    }
-
-    /// Resolve edges scoped to dirty files only.
-    ///
-    /// Processes: edges originating from dirty files (freshly extracted)
-    /// and edges whose target was just invalidated (target_id set to NULL).
-    /// Uses the same 6-tier heuristic as `resolve_edges`.
-    /// Resolve edges after scoped invalidation.
-    ///
-    /// After `invalidate_edges_targeting` has cleared target_ids for edges
-    /// pointing into dirty files, this re-resolves all currently unresolved edges.
-    /// Fewer edges are unresolved compared to a first-time full resolve.
-    pub fn resolve_edges_scoped(
-        &self,
-        dirty_files: &std::collections::HashSet<String>,
-    ) -> Result<u32> {
-        let tx = self.conn.unchecked_transaction()?;
-        let total = self.resolve_edges_scoped_in_tx(dirty_files)?;
-        tx.commit()?;
-        Ok(total)
-    }
-
-    /// Like [`Self::resolve_edges_scoped`] but assumes the caller already
-    /// holds an open transaction.
-    pub fn resolve_edges_scoped_in_tx(
-        &self,
-        dirty_files: &std::collections::HashSet<String>,
-    ) -> Result<u32> {
-        if dirty_files.is_empty() {
-            return Ok(0);
-        }
-        // After invalidation, the set of unresolved edges is naturally scoped:
-        // only edges from dirty files (freshly extracted) or targeting dirty files
-        // (just invalidated) have target_id = NULL.
-        // Reuse the same 2-pass resolution.
-        self.resolve_edges_in_tx()
-    }
-
-    /// Recompute in-degree centrality only for symbols in/around dirty files.
-    ///
-    /// tx-safe: every internal statement participates in any active outer
-    /// transaction — see [`Self::begin_indexing_tx`]. Does NOT open one of
-    /// its own, unlike the batched `*_in_tx` helpers; outside an outer
-    /// transaction the per-file resets are not atomic with the recompute.
-    pub fn compute_in_degrees_scoped(
-        &self,
-        dirty_files: &std::collections::HashSet<String>,
-    ) -> Result<u32> {
-        if dirty_files.is_empty() {
-            return Ok(0);
-        }
-
-        // Reset in-degree for symbols in dirty files
-        for file in dirty_files {
-            self.conn.execute(
-                "UPDATE symbols SET in_degree = 0 WHERE file_path = ?1",
-                params![file],
-            )?;
-        }
-
-        // Also reset symbols that are targets of edges from dirty files
-        // (their in-degree may have changed)
-        for file in dirty_files {
-            self.conn.execute(
-                "UPDATE symbols SET in_degree = 0
-                 WHERE id IN (
-                     SELECT DISTINCT e.target_id FROM edges e
-                     WHERE e.file_path = ?1 AND e.target_id IS NOT NULL
-                 )",
-                params![file],
-            )?;
-        }
-
-        // Recompute for all symbols with in_degree = 0 that have incoming edges
-        let updated = self.conn.execute(
-            "WITH counts AS (
-                SELECT target_id, COUNT(*) AS cnt
-                FROM edges WHERE target_id IS NOT NULL
-                GROUP BY target_id
-            )
-            UPDATE symbols SET in_degree = (
-                SELECT cnt FROM counts WHERE counts.target_id = symbols.id
-            )
-            WHERE in_degree = 0
-              AND id IN (SELECT target_id FROM counts)",
-            [],
-        )?;
-
-        Ok(updated as u32)
-    }
-
-    // ── Queries ──
-
-    /// Search for symbols by name — case-insensitive, prefix match ranks before substring.
-    ///
-    /// `%` and `_` in `query` are treated as literals, not LIKE wildcards.
-    /// Note: `LOWER()` in SQLite is ASCII-only, which is acceptable for code identifiers.
-    /// Returns an error if `query` is empty or `limit` is zero.
-    pub fn search(
-        &self,
-        query: &str,
-        kind_filter: Option<SymbolKind>,
-        file_filter: Option<&str>,
-        limit: u32,
-    ) -> Result<Vec<Symbol>> {
-        anyhow::ensure!(!query.is_empty(), "search query cannot be empty");
-        anyhow::ensure!(limit > 0, "search limit must be at least 1");
-
-        // Escape LIKE special characters so query is matched literally.
-        let escaped = query
-            .replace('\\', "\\\\")
-            .replace('%', "\\%")
-            .replace('_', "\\_");
-        let kind_str = kind_filter.map(|k| k.as_str());
-        // Ranking: match_tier + kind_penalty.
-        //   match_tier: 0 = exact, 1 = prefix, 2 = substring
-        //   kind_penalty: definitions (function/method/class) = 0, variable = 3, import = 6
-        // Definitions always rank above variables/imports across all match tiers:
-        //   exact class=0, prefix function=1, substring method=2,
-        //   exact variable=3, prefix variable=4, substring variable=5,
-        //   exact import=6, ...
-        // Within the same rank score, secondary sort by kind (fn < method < class)
-        // then by file_path and start_line for determinism.
-        let mut stmt = self.conn.prepare(
-            "SELECT id, name, kind, file_path, start_line, end_line,
-                    start_byte, end_byte, parent_id, signature, visibility,
-                    is_async, docstring, in_degree, content_hash, subtree_hash,
-                    (CASE
-                       WHEN LOWER(name) = LOWER(?1)                    THEN 0
-                       WHEN LOWER(name) LIKE LOWER(?2) || '%' ESCAPE '\\' THEN 1
-                       ELSE                                                  2
-                     END) +
-                    (CASE kind
-                       WHEN 'function' THEN 0
-                       WHEN 'method'   THEN 0
-                       WHEN 'class'    THEN 0
-                       WHEN 'variable' THEN 3
-                       WHEN 'import'   THEN 6
-                       ELSE                 3
-                     END) AS rank
-             FROM symbols
-             WHERE LOWER(name) LIKE '%' || LOWER(?2) || '%' ESCAPE '\\'
-               AND (?3 IS NULL OR kind = ?3)
-               AND (?4 IS NULL OR file_path = ?4)
-             ORDER BY rank,
-                      in_degree DESC,
-                      CASE kind
-                        WHEN 'function' THEN 0
-                        WHEN 'method'   THEN 1
-                        WHEN 'class'    THEN 2
-                        ELSE                 3
-                      END,
-                      file_path, start_line
-             LIMIT ?5",
-        )?;
-        // ?1 = raw query (exact equality), ?2 = escaped query (LIKE patterns), ?3 = kind, ?4 = file, ?5 = limit
-        let rows = stmt
-            .query_map(
-                params![query, escaped, kind_str, file_filter, limit],
-                row_to_symbol,
-            )?
-            .collect::<std::result::Result<Vec<_>, _>>()?;
-        Ok(rows)
-    }
-
-    /// Outline: all symbols in a file, ordered by line.
-    pub fn outline(&self, file_path: &str) -> Result<Vec<Symbol>> {
-        let mut stmt = self.conn.prepare(
-            "SELECT id, name, kind, file_path, start_line, end_line, start_byte, end_byte,
-                    parent_id, signature, visibility, is_async, docstring, in_degree,
-                    content_hash, subtree_hash
-             FROM symbols WHERE file_path = ?1
-             ORDER BY start_line",
-        )?;
-        let rows = stmt
-            .query_map(params![file_path], row_to_symbol)?
-            .collect::<std::result::Result<Vec<_>, _>>()?;
-        Ok(rows)
-    }
-
-    /// Find what a symbol calls (edges originating from symbols matching the name).
-    pub fn callees(&self, name: &str) -> Result<Vec<Edge>> {
-        let mut stmt = self.conn.prepare(
-            "SELECT e.id, e.source_id, e.target_name, e.target_id, e.kind, e.file_path, e.line
-             FROM edges e
-             JOIN symbols s ON e.source_id = s.id
-             WHERE s.name = ?1 AND e.kind = 'calls'",
-        )?;
-        let rows = stmt
-            .query_map(params![name], row_to_edge)?
-            .collect::<std::result::Result<Vec<_>, _>>()?;
-        Ok(rows)
-    }
-
-    /// All references to a name, with the source symbol resolved.
-    /// Optionally filter by edge kind.
-    pub fn refs(
-        &self,
-        name: &str,
-        kind_filter: Option<EdgeKind>,
-    ) -> Result<Vec<(Edge, Option<Symbol>)>> {
-        // Use a LEFT JOIN to resolve target_id → symbol name instead of a correlated subquery.
-        let map_row = |row: &rusqlite::Row<'_>| -> rusqlite::Result<(Edge, Option<Symbol>)> {
-            let kind_str = row.get::<_, String>(4)?;
-            let kind = kind_str.parse().unwrap_or(EdgeKind::References);
-            let edge = Edge {
-                source_id: row.get(1)?,
-                target_name: row.get(2)?,
-                target_id: row.get(3)?,
-                kind,
-                file_path: row.get(5)?,
-                line: row.get(6)?,
-            };
-            let sym: Option<Symbol> = if row.get::<_, Option<String>>(7)?.is_some() {
-                Some(row_to_symbol_offset(row, 7)?)
-            } else {
-                None
-            };
-            Ok((edge, sym))
-        };
-
-        let rows = if let Some(kind) = kind_filter {
-            let mut stmt = self.conn.prepare_cached(
-                "SELECT e.id, e.source_id, e.target_name, e.target_id, e.kind, e.file_path, e.line,
-                        s.id, s.name, s.kind, s.file_path, s.start_line, s.end_line,
-                        s.start_byte, s.end_byte, s.parent_id, s.signature, s.visibility,
-                        s.is_async, s.docstring, s.in_degree, s.content_hash, s.subtree_hash
-                 FROM edges e
-                 LEFT JOIN symbols s ON e.source_id = s.id
-                 LEFT JOIN symbols sym2 ON e.target_id = sym2.id
-                 WHERE (e.target_name = ?1 OR sym2.name = ?1)
-                   AND e.kind = ?2",
-            )?;
-            let rows = stmt
-                .query_map(params![name, kind.as_str()], map_row)?
-                .collect::<std::result::Result<Vec<_>, _>>()?;
-            rows
-        } else {
-            let mut stmt = self.conn.prepare_cached(
-                "SELECT e.id, e.source_id, e.target_name, e.target_id, e.kind, e.file_path, e.line,
-                        s.id, s.name, s.kind, s.file_path, s.start_line, s.end_line,
-                        s.start_byte, s.end_byte, s.parent_id, s.signature, s.visibility,
-                        s.is_async, s.docstring, s.in_degree, s.content_hash, s.subtree_hash
-                 FROM edges e
-                 LEFT JOIN symbols s ON e.source_id = s.id
-                 LEFT JOIN symbols sym2 ON e.target_id = sym2.id
-                 WHERE e.target_name = ?1 OR sym2.name = ?1",
-            )?;
-            let rows = stmt
-                .query_map(params![name], map_row)?
-                .collect::<std::result::Result<Vec<_>, _>>()?;
-            rows
-        };
-        Ok(rows)
-    }
-
-    /// Inheritance hierarchy rooted at a class.
-    pub fn hierarchy(&self, class_name: &str) -> Result<Vec<(String, String)>> {
-        // Returns (child, parent) pairs
-        let mut stmt = self.conn.prepare(
-            "SELECT s.name, e.target_name
-             FROM edges e
-             JOIN symbols s ON e.source_id = s.id
-             WHERE e.kind = 'inherits'
-               AND (s.name = ?1 OR e.target_name = ?1)",
-        )?;
-        let rows = stmt
-            .query_map(params![class_name], |row| Ok((row.get(0)?, row.get(1)?)))?
-            .collect::<std::result::Result<Vec<_>, _>>()?;
-        Ok(rows)
-    }
-
-    /// File-level dependencies (imports from a file).
-    pub fn file_deps(&self, file_path: &str) -> Result<Vec<Edge>> {
-        let mut stmt = self.conn.prepare(
-            "SELECT e.id, e.source_id, e.target_name, e.target_id, e.kind, e.file_path, e.line
-             FROM edges e
-             WHERE e.file_path = ?1 AND e.kind = 'imports'",
-        )?;
-        let rows = stmt
-            .query_map(params![file_path], row_to_edge)?
-            .collect::<std::result::Result<Vec<_>, _>>()?;
-        Ok(rows)
-    }
-
-    /// Transitive impact analysis: everything reachable within `depth` hops.
-    ///
-    /// Evaluated as a single recursive CTE rather than iterating `refs()` per
-    /// frontier node — saves N round-trips and lets SQLite's planner amortize
-    /// the LEFT JOINs. Each unique edge is returned once, labeled with the
-    /// minimum depth at which it was reached.
-    pub fn impact(&self, name: &str, max_depth: u32) -> Result<Vec<(Edge, u32)>> {
-        if max_depth == 0 {
-            return Ok(Vec::new());
-        }
-
-        let sql = "
-            WITH RECURSIVE impacted(
-                edge_id, source_id, target_name, target_id, kind,
-                file_path, line, source_name, depth
-            ) AS (
-                SELECT e.id, e.source_id, e.target_name, e.target_id, e.kind,
-                       e.file_path, e.line, s.name, 1
-                FROM edges e
-                LEFT JOIN symbols s ON e.source_id = s.id
-                LEFT JOIN symbols sym2 ON e.target_id = sym2.id
-                WHERE e.target_name = ?1 OR sym2.name = ?1
-
-                UNION
-
-                SELECT e.id, e.source_id, e.target_name, e.target_id, e.kind,
-                       e.file_path, e.line, s.name, i.depth + 1
-                FROM impacted i
-                JOIN edges e
-                  ON (e.target_name = i.source_name
-                      OR EXISTS (
-                          SELECT 1 FROM symbols t
-                          WHERE t.id = e.target_id AND t.name = i.source_name
-                      ))
-                LEFT JOIN symbols s ON e.source_id = s.id
-                WHERE i.source_name IS NOT NULL AND i.depth < ?2
-            )
-            SELECT source_id, target_name, target_id, kind, file_path, line,
-                   MIN(depth) AS depth
-            FROM impacted
-            GROUP BY edge_id
-            ORDER BY depth, edge_id
-        ";
-
-        let mut stmt = self.conn.prepare_cached(sql)?;
-        let rows = stmt
-            .query_map(params![name, max_depth], |row| {
-                let kind_str: String = row.get(3)?;
-                let kind = kind_str.parse().unwrap_or(EdgeKind::References);
-                let edge = Edge {
-                    source_id: row.get(0)?,
-                    target_name: row.get(1)?,
-                    target_id: row.get(2)?,
-                    kind,
-                    file_path: row.get(4)?,
-                    line: row.get(5)?,
-                };
-                let depth: u32 = row.get(6)?;
-                Ok((edge, depth))
-            })?
-            .collect::<std::result::Result<Vec<_>, _>>()?;
-        Ok(rows)
-    }
-
-    /// True when no symbols have been indexed yet (fresh/empty DB). Cheap —
-    /// a single `EXISTS(SELECT 1 FROM symbols)` that stops at the first row.
-    /// Used by query commands to distinguish "no index yet" from a no-match.
-    pub fn is_empty(&self) -> Result<bool> {
-        let exists: bool =
-            self.conn
-                .query_row("SELECT EXISTS(SELECT 1 FROM symbols)", [], |row| row.get(0))?;
-        Ok(!exists)
-    }
-
-    /// Index statistics.
-    pub fn stats(&self) -> Result<IndexStats> {
-        let num_files: u32 = self
-            .conn
-            .query_row("SELECT COUNT(*) FROM files", [], |row| row.get(0))?;
-        let num_symbols: u32 = self
-            .conn
-            .query_row("SELECT COUNT(*) FROM symbols", [], |row| row.get(0))?;
-        // Single GROUP BY over edges replaces what would otherwise be four
-        // sequential full table scans (one COUNT(*) per bucket). The partial
-        // index `idx_edges_unresolved` only covers state=0, so state=2 and
-        // state=3 counts can't use it — one scan + a 4-row Vec is cheaper.
-        let mut bucket_stmt = self
-            .conn
-            .prepare("SELECT resolution_state, COUNT(*) FROM edges GROUP BY resolution_state")?;
-        let mut num_resolved: u32 = 0;
-        let mut num_unresolvable: u32 = 0;
-        let mut num_external: u32 = 0;
-        let mut num_edges: u32 = 0;
-        let rows = bucket_stmt.query_map([], |row| {
-            let state: i64 = row.get(0)?;
-            let count: u32 = row.get(1)?;
-            Ok((state, count))
-        })?;
-        for row in rows {
-            let (state, count) = row?;
-            num_edges += count;
-            match state {
-                1 => num_resolved = count,
-                2 => num_unresolvable = count,
-                3 => num_external = count,
-                _ => {} // state=0 (unresolved) or any future state
-            }
-        }
-
-        let mut lang_stmt = self.conn.prepare(
-            "SELECT language, COUNT(*) FROM files GROUP BY language ORDER BY COUNT(*) DESC",
-        )?;
-        let languages: Vec<(String, u32)> = lang_stmt
-            .query_map([], |row| Ok((row.get(0)?, row.get(1)?)))?
-            .collect::<std::result::Result<Vec<_>, _>>()?;
-
-        let mut kind_stmt = self
-            .conn
-            .prepare("SELECT kind, COUNT(*) FROM symbols GROUP BY kind ORDER BY COUNT(*) DESC")?;
-        let symbol_kinds: Vec<(String, u32)> = kind_stmt
-            .query_map([], |row| Ok((row.get(0)?, row.get(1)?)))?
-            .collect::<std::result::Result<Vec<_>, _>>()?;
-
-        Ok(IndexStats {
-            num_files,
-            num_symbols,
-            num_edges,
-            num_resolved,
-            num_unresolvable,
-            num_external,
-            languages,
-            symbol_kinds,
-        })
-    }
-
-    /// Record a successful query against the index for the `cartog stats --savings`
-    /// / `cartog savings` retention hook.
-    ///
-    /// Best-effort: errors are swallowed (logged via `warn!`) so a failing
-    /// write never aborts the user's actual query.
-    ///
-    /// **Read-only attach skips the write.** Secondary MCP servers opened via
-    /// [`Self::open_readonly`] cannot write at all. As a result, queries
-    /// served by a secondary are NOT reflected in `query_log` — there is no
-    /// IPC that forwards them to the primary. `cartog stats --savings` on a
-    /// machine that runs multiple MCP servers will therefore *undercount*
-    /// secondary traffic, not overcount it. This is a deliberate trade-off:
-    /// the alternative would be a separate per-process file with its own
-    /// merge logic, which is more complexity than the retention hook needs.
-    ///
-    /// Stored fields: `tool` (e.g. `"search"`, `"refs"`, MCP-side already
-    /// strips the `cartog_` prefix so CLI and MCP rows aggregate), `source`
-    /// (`"cli"` or `"mcp"`), and a unix-seconds timestamp. The query payload
-    /// itself is never recorded — see the privacy banner in README.
-    pub fn log_query(&self, tool: &str, source: &str) {
-        if self.is_read_only() {
-            return;
-        }
-        let ts = std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .map(|d| d.as_secs() as i64)
-            .unwrap_or(0);
-        if let Err(e) = self.conn.execute(
-            "INSERT INTO query_log (tool, source, ts) VALUES (?1, ?2, ?3)",
-            params![tool, source, ts],
-        ) {
-            // Always warn for the individual failure (debuggable in traces),
-            // and additionally emit a one-shot loud-error on the first
-            // failure so a persistently-broken query_log (e.g. SQLITE_FULL
-            // or a missing table on a manually-tampered DB) is visible even
-            // when warns are filtered.
-            warn!(error = %e, tool, source, "query_log insert failed");
-            if !LOG_QUERY_FAILURE_REPORTED.swap(true, std::sync::atomic::Ordering::Relaxed) {
-                tracing::error!(
-                    error = %e,
-                    "query_log is failing — `cartog stats --savings` will undercount. \
-                     Check disk space and `cartog doctor`."
-                );
-            }
-        }
-    }
-
-    /// Aggregate `query_log` for `cartog stats --savings` / `cartog savings`.
-    /// Safe on read-only attach (it's a read). Returns an empty report when
-    /// the `query_log` table is missing (the read-only attach path skips
-    /// schema bootstrap, so a v5 DB that lost the table — manual drop, partial
-    /// snapshot restore — would otherwise surface a `no such table` error).
-    pub fn savings_breakdown(&self) -> Result<SavingsReport> {
-        // Probe for the table once. Only treat "no such table" as the empty-
-        // report case so real DB faults (corruption, locked, permissions)
-        // still propagate to the caller.
-        if let Err(e) = self.conn.prepare("SELECT 1 FROM query_log LIMIT 0") {
-            if is_no_such_table(&e) {
-                return Ok(empty_savings_report());
-            }
-            return Err(e.into());
-        }
-
-        let mut tool_stmt = self.conn.prepare(
-            "SELECT tool, COUNT(*) FROM query_log GROUP BY tool ORDER BY COUNT(*) DESC, tool",
-        )?;
-        let by_tool: Vec<(String, u64)> = tool_stmt
-            .query_map([], |row| Ok((row.get(0)?, row.get::<_, i64>(1)? as u64)))?
-            .collect::<std::result::Result<Vec<_>, _>>()?;
-
-        let mut src_stmt = self.conn.prepare(
-            "SELECT source, COUNT(*) FROM query_log GROUP BY source ORDER BY COUNT(*) DESC, source",
-        )?;
-        let by_source: Vec<(String, u64)> = src_stmt
-            .query_map([], |row| Ok((row.get(0)?, row.get::<_, i64>(1)? as u64)))?
-            .collect::<std::result::Result<Vec<_>, _>>()?;
-
-        let total_queries: u64 = by_tool.iter().map(|(_, c)| c).sum();
-        let tokens_used_cartog = total_queries.saturating_mul(TOKENS_PER_QUERY_CARTOG as u64);
-        let tokens_used_grep = total_queries.saturating_mul(TOKENS_PER_QUERY_GREP as u64);
-        let estimated_tokens_saved = tokens_used_grep.saturating_sub(tokens_used_cartog);
-        // 0–100, computed with saturating_mul to match the rest of the
-        // arithmetic in this function (everything else uses saturating_* to
-        // stay safe at extreme query counts).
-        let percent_saved = estimated_tokens_saved
-            .saturating_mul(100)
-            .checked_div(tokens_used_grep)
-            .unwrap_or(0)
-            .min(100) as u8;
-
-        Ok(SavingsReport {
-            by_tool,
-            by_source,
-            total_queries,
-            tokens_used_cartog,
-            tokens_used_grep,
-            estimated_tokens_saved,
-            percent_saved,
-            baseline_delta: TOKENS_SAVED_PER_QUERY,
-        })
-    }
-
-    /// Get all non-import symbols ordered by in-degree (highest first), then by file.
-    ///
-    /// Used by `cartog map` to produce a centrality-ranked codebase summary.
-    pub fn top_symbols(&self, limit: u32) -> Result<Vec<Symbol>> {
-        let mut stmt = self.conn.prepare(
-            "SELECT id, name, kind, file_path, start_line, end_line, start_byte, end_byte,
-                    parent_id, signature, visibility, is_async, docstring, in_degree,
-                    content_hash, subtree_hash
-             FROM symbols
-             WHERE kind != 'import' AND kind != 'variable'
-             ORDER BY in_degree DESC, file_path, start_line
-             LIMIT ?1",
-        )?;
-        let rows = stmt
-            .query_map(params![limit], row_to_symbol)?
-            .collect::<std::result::Result<Vec<_>, _>>()?;
-        Ok(rows)
-    }
-
-    /// Returns `true` if at least one file has been indexed.
-    ///
-    /// Cheaper than [`Database::stats`] for the common "is the index empty?" check —
-    /// SQLite can satisfy `LIMIT 1` with a single index seek rather than a full count.
-    pub fn has_indexed_files(&self) -> Result<bool> {
-        Ok(self
-            .conn
-            .query_row("SELECT 1 FROM files LIMIT 1", [], |_| Ok(()))
-            .optional()?
-            .is_some())
-    }
-
-    /// Get symbols for a set of file paths, grouped by file, ordered by line.
-    ///
-    /// Optionally filter by symbol kind. Only returns symbols for files that
-    /// exist in the index. Files with no matching symbols are omitted.
-    /// SQLite variable limit per query. Chunking keeps us well under the default
-    /// `SQLITE_MAX_VARIABLE_NUMBER` (999 in older builds, 32766 in newer).
-    const FILE_CHUNK_SIZE: usize = 500;
-
-    pub fn symbols_for_files(
-        &self,
-        file_paths: &[String],
-        kind_filter: Option<SymbolKind>,
-    ) -> Result<Vec<Symbol>> {
-        if file_paths.is_empty() {
-            return Ok(Vec::new());
-        }
-
-        let kind_str = kind_filter.map(|k| k.as_str().to_string());
-        let mut all_results = Vec::new();
-
-        for chunk in file_paths.chunks(Self::FILE_CHUNK_SIZE) {
-            let placeholders: Vec<_> = (1..=chunk.len()).map(|i| format!("?{i}")).collect();
-            let kind_param_idx = chunk.len() + 1;
-
-            let sql = format!(
-                "SELECT id, name, kind, file_path, start_line, end_line, start_byte, end_byte,
-                        parent_id, signature, visibility, is_async, docstring, in_degree,
-                    content_hash, subtree_hash
-                 FROM symbols
-                 WHERE file_path IN ({})
-                   AND (?{kind_param_idx} IS NULL OR kind = ?{kind_param_idx})
-                 ORDER BY file_path, start_line",
-                placeholders.join(", ")
-            );
-            let mut stmt = self.conn.prepare(&sql)?;
-
-            let mut param_values: Vec<Box<dyn rusqlite::types::ToSql>> = chunk
-                .iter()
-                .map(|p| Box::new(p.clone()) as Box<dyn rusqlite::types::ToSql>)
-                .collect();
-            param_values.push(Box::new(kind_str.clone()));
-
-            let params: Vec<&dyn rusqlite::types::ToSql> =
-                param_values.iter().map(|p| &**p).collect();
-            let rows = stmt
-                .query_map(&*params, row_to_symbol)?
-                .collect::<std::result::Result<Vec<_>, _>>()?;
-            all_results.extend(rows);
-        }
-
-        // Re-sort across chunks to maintain file_path, start_line order
-        if file_paths.len() > Self::FILE_CHUNK_SIZE {
-            all_results.sort_by(|a, b| {
-                a.file_path
-                    .cmp(&b.file_path)
-                    .then(a.start_line.cmp(&b.start_line))
-            });
-        }
-
-        Ok(all_results)
-    }
-
-    /// Get all indexed file paths, sorted alphabetically.
-    pub fn all_files(&self) -> Result<Vec<String>> {
-        let mut stmt = self.conn.prepare("SELECT path FROM files ORDER BY path")?;
-        let rows = stmt
-            .query_map([], |row| row.get(0))?
-            .collect::<std::result::Result<Vec<_>, _>>()?;
-        Ok(rows)
-    }
-
-    /// Load (path, hash) pairs for every indexed file in one query.
-    ///
-    /// Used by the parallel indexer to avoid per-file DB round trips when
-    /// deciding whether a file needs re-parsing.
-    pub fn all_file_hashes(&self) -> Result<std::collections::HashMap<String, String>> {
-        let mut stmt = self.conn.prepare("SELECT path, hash FROM files")?;
-        let rows = stmt
-            .query_map([], |row| {
-                Ok((row.get::<_, String>(0)?, row.get::<_, String>(1)?))
-            })?
-            .collect::<std::result::Result<Vec<_>, _>>()?;
-        Ok(rows.into_iter().collect())
-    }
-
-    // ── RAG: Symbol Content ──
-
-    /// Insert or replace symbol content (raw source + metadata header for embedding).
-    ///
-    /// `symbol_name` is used to compute a normalized form (camelCase/snake_case split)
-    /// stored in the FTS5 index for better keyword matching.
-    pub fn upsert_symbol_content(
-        &self,
-        symbol_id: &str,
-        symbol_name: &str,
-        content: &str,
-        header: &str,
-    ) -> Result<()> {
-        let normalized = normalize_symbol_name(symbol_name);
-        self.conn.execute(
-            "INSERT OR REPLACE INTO symbol_content (symbol_id, content, header, normalized_name)
-             VALUES (?1, ?2, ?3, ?4)",
-            params![symbol_id, content, header, normalized],
-        )?;
-        Ok(())
-    }
-
-    /// Insert multiple symbol contents in a single transaction.
-    ///
-    /// Tuples: `(symbol_id, symbol_name, content, header)`.
-    pub fn insert_symbol_contents(&self, items: &[(String, String, String, String)]) -> Result<()> {
-        let tx = self.conn.unchecked_transaction()?;
-        self.insert_symbol_contents_in_tx(items)?;
-        tx.commit()?;
-        Ok(())
-    }
-
-    /// Like [`Self::insert_symbol_contents`] but assumes the caller already
-    /// holds an open transaction.
-    pub fn insert_symbol_contents_in_tx(
-        &self,
-        items: &[(String, String, String, String)],
-    ) -> Result<()> {
-        let mut stmt = self.conn.prepare_cached(
-            "INSERT OR REPLACE INTO symbol_content (symbol_id, content, header, normalized_name)
-             VALUES (?1, ?2, ?3, ?4)",
-        )?;
-        for (symbol_id, name, content, header) in items {
-            let normalized = normalize_symbol_name(name);
-            stmt.execute(params![symbol_id, content, header, normalized])?;
-        }
-        Ok(())
-    }
-
-    /// Remove symbol content for all symbols in a file.
-    pub fn clear_symbol_content_for_file(&self, file_path: &str) -> Result<()> {
-        self.conn.execute(
-            "DELETE FROM symbol_content WHERE symbol_id IN
-             (SELECT id FROM symbols WHERE file_path = ?1)",
-            params![file_path],
-        )?;
-        Ok(())
-    }
-
-    /// Get the content + header for a symbol.
-    pub fn get_symbol_content(&self, symbol_id: &str) -> Result<Option<(String, String)>> {
-        self.conn
-            .query_row(
-                "SELECT content, header FROM symbol_content WHERE symbol_id = ?1",
-                params![symbol_id],
-                |row| Ok((row.get(0)?, row.get(1)?)),
-            )
-            .optional()
-            .context("Failed to query symbol content")
-    }
-
-    /// Batch fetch content + header for multiple symbols.
-    ///
-    /// Returns a map of `symbol_id → (content, header)` for all found symbols.
-    pub fn get_symbol_contents_batch(
-        &self,
-        symbol_ids: &[String],
-    ) -> Result<std::collections::HashMap<String, (String, String)>> {
-        let mut result = std::collections::HashMap::with_capacity(symbol_ids.len());
-        if symbol_ids.is_empty() {
-            return Ok(result);
-        }
-        for chunk in symbol_ids.chunks(Self::FILE_CHUNK_SIZE) {
-            let placeholders: Vec<&str> = chunk.iter().map(|_| "?").collect();
-            let sql = format!(
-                "SELECT symbol_id, content, header FROM symbol_content WHERE symbol_id IN ({})",
-                placeholders.join(",")
-            );
-            let mut stmt = self.conn.prepare(&sql)?;
-            let params: Vec<Box<dyn rusqlite::types::ToSql>> = chunk
-                .iter()
-                .map(|id| Box::new(id.clone()) as Box<dyn rusqlite::types::ToSql>)
-                .collect();
-            let param_refs: Vec<&dyn rusqlite::types::ToSql> =
-                params.iter().map(|p| p.as_ref()).collect();
-            let rows = stmt
-                .query_map(param_refs.as_slice(), |row| {
-                    Ok((
-                        row.get::<_, String>(0)?,
-                        row.get::<_, String>(1)?,
-                        row.get::<_, String>(2)?,
-                    ))
-                })?
-                .collect::<std::result::Result<Vec<_>, _>>()?;
-            for (id, content, header) in rows {
-                result.insert(id, (content, header));
-            }
-        }
-        Ok(result)
-    }
-
-    // ── RAG: FTS5 Search ──
-
-    /// Full-text search over symbol names and content using BM25 ranking.
-    ///
-    /// Returns symbol IDs ordered by relevance (best match first).
-    pub fn fts5_search(&self, query: &str, limit: u32) -> Result<Vec<String>> {
-        let mut stmt = self.conn.prepare(
-            "SELECT sc.symbol_id
-             FROM symbol_fts f
-             JOIN symbol_content sc ON sc.rowid = f.rowid
-             WHERE symbol_fts MATCH ?1
-             ORDER BY rank
-             LIMIT ?2",
-        )?;
-        let rows = stmt
-            .query_map(params![query, limit], |row| row.get(0))?
-            .collect::<std::result::Result<Vec<_>, _>>()?;
-        Ok(rows)
-    }
-
-    // ── RAG: Embedding Map ──
-
-    /// Get or create an integer ID for a symbol in the embedding map.
-    ///
-    /// Returns the `id` (integer rowid) used as key in the vec0 virtual table.
-    pub fn get_or_create_embedding_id(&self, symbol_id: &str) -> Result<i64> {
-        // Try to get existing
-        let existing: Option<i64> = self
-            .conn
-            .query_row(
-                "SELECT id FROM symbol_embedding_map WHERE symbol_id = ?1",
-                params![symbol_id],
-                |row| row.get(0),
-            )
-            .optional()?;
-
-        if let Some(id) = existing {
-            return Ok(id);
-        }
-
-        // Insert new
-        self.conn.execute(
-            "INSERT INTO symbol_embedding_map (symbol_id) VALUES (?1)",
-            params![symbol_id],
-        )?;
-        Ok(self.conn.last_insert_rowid())
-    }
-
-    /// Look up the symbol ID for an embedding map rowid.
-    pub fn symbol_id_for_embedding(&self, embedding_id: i64) -> Result<Option<String>> {
-        self.conn
-            .query_row(
-                "SELECT symbol_id FROM symbol_embedding_map WHERE id = ?1",
-                params![embedding_id],
-                |row| row.get(0),
-            )
-            .optional()
-            .context("Failed to query embedding map")
-    }
-
-    /// Batch look up symbol IDs for multiple embedding map rowids.
-    pub fn symbol_ids_for_embeddings(&self, embedding_ids: &[i64]) -> Result<Vec<(i64, String)>> {
-        if embedding_ids.is_empty() {
-            return Ok(Vec::new());
-        }
-        let mut all_results = Vec::with_capacity(embedding_ids.len());
-        for chunk in embedding_ids.chunks(Self::FILE_CHUNK_SIZE) {
-            let placeholders: Vec<String> = chunk.iter().map(|_| "?".to_string()).collect();
-            let sql = format!(
-                "SELECT id, symbol_id FROM symbol_embedding_map WHERE id IN ({})",
-                placeholders.join(",")
-            );
-            let mut stmt = self.conn.prepare(&sql)?;
-            let params: Vec<Box<dyn rusqlite::types::ToSql>> = chunk
-                .iter()
-                .map(|id| Box::new(*id) as Box<dyn rusqlite::types::ToSql>)
-                .collect();
-            let param_refs: Vec<&dyn rusqlite::types::ToSql> =
-                params.iter().map(|p| p.as_ref()).collect();
-            let rows = stmt
-                .query_map(param_refs.as_slice(), |row| Ok((row.get(0)?, row.get(1)?)))?
-                .collect::<std::result::Result<Vec<_>, _>>()?;
-            all_results.extend(rows);
-        }
-        Ok(all_results)
-    }
-
-    // ── RAG: Vector Storage (sqlite-vec) ──
-
-    /// Insert or replace an embedding vector for a symbol.
-    ///
-    /// `embedding_id` is the integer key from `symbol_embedding_map`.
-    /// `embedding` is a 384-dim f32 vector serialized as little-endian bytes.
-    pub fn upsert_embedding(&self, embedding_id: i64, embedding: &[u8]) -> Result<()> {
-        // Delete existing entry if any (vec0 doesn't support REPLACE)
-        self.conn.execute(
-            "DELETE FROM symbol_vec WHERE rowid = ?1",
-            params![embedding_id],
-        )?;
-        self.conn.execute(
-            "INSERT INTO symbol_vec (rowid, embedding) VALUES (?1, ?2)",
-            params![embedding_id, embedding],
-        )?;
-        Ok(())
-    }
-
-    /// Insert multiple embeddings in a single transaction.
-    pub fn insert_embeddings(&self, items: &[(i64, Vec<u8>)]) -> Result<()> {
-        let tx = self.conn.unchecked_transaction()?;
-        for (id, embedding) in items {
-            self.conn
-                .execute("DELETE FROM symbol_vec WHERE rowid = ?1", params![id])?;
-            self.conn.execute(
-                "INSERT INTO symbol_vec (rowid, embedding) VALUES (?1, ?2)",
-                params![id, embedding],
-            )?;
-        }
-        tx.commit()?;
-        Ok(())
-    }
-
-    /// KNN vector search: find the `limit` nearest neighbors to `query_embedding`.
-    ///
-    /// Returns `(embedding_id, distance)` pairs ordered by distance (ascending).
-    pub fn vector_search(&self, query_embedding: &[u8], limit: u32) -> Result<Vec<(i64, f64)>> {
-        let mut stmt = self.conn.prepare(
-            "SELECT rowid, distance
-             FROM symbol_vec
-             WHERE embedding MATCH ?1
-             ORDER BY distance
-             LIMIT ?2",
-        )?;
-        let rows = stmt
-            .query_map(params![query_embedding, limit], |row| {
-                Ok((row.get(0)?, row.get(1)?))
-            })?
-            .collect::<std::result::Result<Vec<_>, _>>()?;
-        Ok(rows)
-    }
-
-    /// Count the number of embeddings stored.
-    pub fn embedding_count(&self) -> Result<u32> {
-        Ok(self
-            .conn
-            .query_row("SELECT COUNT(*) FROM symbol_embedding_map", [], |row| {
-                row.get(0)
-            })?)
-    }
-
-    /// Check if a symbol already has an embedding.
-    pub fn has_embedding(&self, symbol_id: &str) -> Result<bool> {
-        let map_id: Option<i64> = self
-            .conn
-            .query_row(
-                "SELECT id FROM symbol_embedding_map WHERE symbol_id = ?1",
-                params![symbol_id],
-                |row| row.get(0),
-            )
-            .optional()?;
-
-        if let Some(id) = map_id {
-            let exists: bool = self.conn.query_row(
-                "SELECT EXISTS(SELECT 1 FROM symbol_vec WHERE rowid = ?1)",
-                params![id],
-                |row| row.get(0),
-            )?;
-            Ok(exists)
-        } else {
-            Ok(false)
-        }
-    }
-
-    /// Remove all RAG data (content, FTS, embeddings, embedding map) for symbols in a file.
-    pub fn clear_rag_data_for_file(&self, file_path: &str) -> Result<()> {
-        // Delete embeddings via the map
-        self.conn.execute(
-            "DELETE FROM symbol_vec WHERE rowid IN
-             (SELECT em.id FROM symbol_embedding_map em
-              JOIN symbols s ON em.symbol_id = s.id
-              WHERE s.file_path = ?1)",
-            params![file_path],
-        )?;
-        // Delete embedding map entries
-        self.conn.execute(
-            "DELETE FROM symbol_embedding_map WHERE symbol_id IN
-             (SELECT id FROM symbols WHERE file_path = ?1)",
-            params![file_path],
-        )?;
-        // Delete content (triggers will clean up FTS)
-        self.clear_symbol_content_for_file(file_path)?;
-        Ok(())
-    }
-
-    /// Get a symbol by its ID.
-    pub fn get_symbol(&self, id: &str) -> Result<Option<Symbol>> {
-        self.conn
-            .query_row(
-                "SELECT id, name, kind, file_path, start_line, end_line, start_byte, end_byte,
-                        parent_id, signature, visibility, is_async, docstring, in_degree,
-                    content_hash, subtree_hash
-                 FROM symbols WHERE id = ?1",
-                params![id],
-                row_to_symbol,
-            )
-            .optional()
-            .context("Failed to query symbol")
-    }
-
-    /// Get multiple symbols by their IDs, preserving order.
-    pub fn get_symbols_by_ids(&self, ids: &[String]) -> Result<Vec<Symbol>> {
-        if ids.is_empty() {
-            return Ok(Vec::new());
-        }
-        let placeholders: Vec<&str> = ids.iter().map(|_| "?").collect();
-        let sql = format!(
-            "SELECT id, name, kind, file_path, start_line, end_line, start_byte, end_byte,
-                    parent_id, signature, visibility, is_async, docstring, in_degree,
-                    content_hash, subtree_hash
-             FROM symbols WHERE id IN ({})",
-            placeholders.join(",")
-        );
-        let mut stmt = self.conn.prepare(&sql)?;
-        let params: Vec<Box<dyn rusqlite::types::ToSql>> = ids
-            .iter()
-            .map(|id| Box::new(id.clone()) as Box<dyn rusqlite::types::ToSql>)
-            .collect();
-        let param_refs: Vec<&dyn rusqlite::types::ToSql> =
-            params.iter().map(|p| p.as_ref()).collect();
-        let rows: std::collections::HashMap<String, Symbol> = stmt
-            .query_map(param_refs.as_slice(), row_to_symbol)?
-            .filter_map(|r| r.ok())
-            .map(|s| (s.id.clone(), s))
-            .collect();
-        // Preserve caller's ordering
-        Ok(ids.iter().filter_map(|id| rows.get(id).cloned()).collect())
-    }
-
-    /// Get all symbol IDs that have content stored but no embedding yet.
-    ///
-    /// Variables are excluded — they are too numerous and low-signal for embedding.
-    pub fn symbols_needing_embeddings(&self) -> Result<Vec<String>> {
-        let mut stmt = self.conn.prepare(
-            "SELECT sc.symbol_id FROM symbol_content sc
-             JOIN symbols s ON s.id = sc.symbol_id
-             WHERE s.kind NOT IN (?1, ?2)
-             AND NOT EXISTS (
-                 SELECT 1 FROM symbol_embedding_map em
-                 JOIN symbol_vec sv ON sv.rowid = em.id
-                 WHERE em.symbol_id = sc.symbol_id
-             )",
-        )?;
-        let rows = stmt
-            .query_map(
-                params![SymbolKind::Variable.as_str(), SymbolKind::Import.as_str(),],
-                |row| row.get(0),
-            )?
-            .collect::<std::result::Result<Vec<_>, _>>()?;
-        Ok(rows)
-    }
-
-    /// Count symbols that have content stored.
-    pub fn symbol_content_count(&self) -> Result<u32> {
-        Ok(self
-            .conn
-            .query_row("SELECT COUNT(*) FROM symbol_content", [], |row| row.get(0))?)
-    }
-
-    /// Get all symbol IDs that have content stored (excluding variables and imports).
-    pub fn all_content_symbol_ids(&self) -> Result<Vec<String>> {
-        let mut stmt = self.conn.prepare(
-            "SELECT sc.symbol_id FROM symbol_content sc
-             JOIN symbols s ON s.id = sc.symbol_id
-             WHERE s.kind NOT IN (?1, ?2)
-             ORDER BY sc.symbol_id",
-        )?;
-        let rows = stmt
-            .query_map(
-                params![SymbolKind::Variable.as_str(), SymbolKind::Import.as_str(),],
-                |row| row.get(0),
-            )?
-            .collect::<std::result::Result<Vec<_>, _>>()?;
-        Ok(rows)
-    }
-
-    /// Clear all embedding data (for force re-embed).
-    pub fn clear_all_embeddings(&self) -> Result<()> {
-        self.conn.execute("DELETE FROM symbol_vec", [])?;
-        self.conn.execute("DELETE FROM symbol_embedding_map", [])?;
-        Ok(())
-    }
-
-    // ── LSP Resolution Helpers ──
-    //
-    // These helpers (`unresolved_edges`, `find_symbol_at_location`,
-    // `update_edge_target`, `mark_edge_unresolvable`, `mark_edge_external`,
-    // `edge_resolution_state`) are called from `cartog-lsp::lsp_resolve_edges`,
-    // which itself runs inside `index_directory`'s outer indexing transaction
-    // (see [`Self::begin_indexing_tx`]). They MUST remain transaction-free —
-    // any future addition of `unchecked_transaction()` here would re-introduce
-    // the Phase 3 atomicity bug at runtime ("cannot start a transaction
-    // within a transaction").
-
-    /// Return edges still waiting for resolution (`resolution_state = 0`).
-    ///
-    /// Edges marked `state = 2` (unresolvable) or `state = 3` (external) are
-    /// excluded so a dirty reindex doesn't re-query the language server for
-    /// edges it already classified. Both are sticky and re-enter the
-    /// unresolved set only via [`Self::reset_unresolvable_for_names`] when a
-    /// matching symbol is added, or [`Self::reset_all_unresolvable`] on
-    /// `--force`. ([`Self::invalidate_edges_targeting`] only touches state=1
-    /// rows because it filters on `target_id IS NOT NULL`, and state {2, 3}
-    /// rows always have `target_id NULL`.)
-    ///
-    /// tx-safe: read-only single statement — see note above the section header.
-    pub fn unresolved_edges(&self) -> Result<Vec<UnresolvedEdge>> {
-        let mut stmt = self.conn.prepare(
-            "SELECT e.id, e.target_name, e.file_path, e.line
-             FROM edges e
-             WHERE e.resolution_state = 0",
-        )?;
-
-        let rows = stmt.query_map([], |row| {
-            Ok(UnresolvedEdge {
-                edge_id: row.get(0)?,
-                target_name: row.get(1)?,
-                file_path: row.get(2)?,
-                line: row.get(3)?,
-            })
-        })?;
-
-        rows.collect::<rusqlite::Result<Vec<_>>>()
-            .map_err(Into::into)
-    }
-
-    /// Find the tightest-enclosing symbol at a given file + line.
-    ///
-    /// tx-safe: read-only single statement — see the LSP-section header note.
-    pub fn find_symbol_at_location(&self, file_path: &str, line: u32) -> Result<Option<String>> {
-        let id: Option<String> = self
-            .conn
-            .query_row(
-                "SELECT id FROM symbols
-                 WHERE file_path = ?1 AND start_line <= ?2 AND end_line >= ?2
-                 ORDER BY (end_line - start_line) ASC
-                 LIMIT 1",
-                params![file_path, line],
-                |row| row.get(0),
-            )
-            .optional()?;
-        Ok(id)
-    }
-
-    /// Update a single edge's target_id and flip it to `resolution_state = 1`.
-    ///
-    /// tx-safe: single statement — see the LSP-section header note. If you
-    /// ever batch this internally with `unchecked_transaction()`, also update
-    /// `index_directory` so it does not call `lsp_resolve_edges` inside its
-    /// outer transaction.
-    pub fn update_edge_target(&self, edge_id: i64, target_id: &str) -> Result<()> {
-        self.conn.execute(
-            "UPDATE edges SET target_id = ?1, resolution_state = 1 WHERE id = ?2",
-            params![target_id, edge_id],
-        )?;
-        Ok(())
-    }
-
-    /// Test-only inspector: returns true when the edge is at `resolution_state = 2`.
-    ///
-    /// Exposed because the column is otherwise crate-private, and downstream
-    /// integration tests (cartog-indexer) need a read-only way to assert the
-    /// marker state without snapshotting raw SQL.
-    pub fn is_edge_unresolvable(&self, edge_id: i64) -> Result<bool> {
-        Ok(self.edge_resolution_state(edge_id)? == 2)
-    }
-
-    /// Test-only inspector: returns the raw `resolution_state` value for an edge.
-    ///
-    /// 0=unresolved, 1=resolved, 2=unresolvable, 3=external.
-    pub fn edge_resolution_state(&self, edge_id: i64) -> Result<i64> {
-        let state: i64 = self.conn.query_row(
-            "SELECT resolution_state FROM edges WHERE id = ?1",
-            params![edge_id],
-            |row| row.get(0),
-        )?;
-        Ok(state)
-    }
-
-    /// Reset every edge at `resolution_state IN (2, 3)` back to `0`. Used by
-    /// `cartog index --force` to honor the "retry everything" contract:
-    /// without this, the heuristic + LSP would still skip permanently-marked
-    /// edges (both unresolvable and external) even on a forced re-index.
-    ///
-    /// tx-safe: single statement — see the LSP-section header note.
-    pub fn reset_all_unresolvable(&self) -> Result<u32> {
-        let n = self.conn.execute(
-            "UPDATE edges SET resolution_state = 0 WHERE resolution_state IN (2, 3)",
-            [],
-        )?;
-        Ok(n as u32)
-    }
-
-    /// Mark a single edge as `resolution_state = 2` (LSP definitively gave up).
-    ///
-    /// Callers MUST only invoke this after a definitive negative answer from
-    /// the language server. Never call from a transient-error branch (server
-    /// crash, didOpen failure, half-loaded warmup) — the marker is sticky
-    /// across runs until [`Self::reset_unresolvable_for_names`] reopens it
-    /// (on a matching new symbol) or [`Self::reset_all_unresolvable`] runs
-    /// (`--force`).
-    ///
-    /// The `WHERE resolution_state = 0` guard preserves the invariant that
-    /// state {2, 3} rows have `target_id IS NULL` — without it an accidental
-    /// call on a state=1 (resolved) edge would silently flip the state while
-    /// keeping the stale target, hiding a corrupted edge from
-    /// [`Self::unresolved_edges`].
-    ///
-    /// tx-safe: single statement — see the LSP-section header note.
-    pub fn mark_edge_unresolvable(&self, edge_id: i64) -> Result<()> {
-        self.conn.execute(
-            "UPDATE edges SET resolution_state = 2 WHERE id = ?1 AND resolution_state = 0",
-            params![edge_id],
-        )?;
-        Ok(())
-    }
-
-    /// Mark a single edge as `resolution_state = 3` (LSP located the target
-    /// outside the indexed root — stdlib, third-party deps, node_modules).
-    ///
-    /// Same stickiness contract as [`Self::mark_edge_unresolvable`]: only call
-    /// after a definitive positive answer naming an out-of-root URI;
-    /// reopened by the same name-keyed and force-reset paths. The
-    /// `WHERE resolution_state = 0` guard preserves the `target_id IS NULL`
-    /// invariant for state=3 rows.
-    ///
-    /// tx-safe: single statement — see the LSP-section header note.
-    pub fn mark_edge_external(&self, edge_id: i64) -> Result<()> {
-        self.conn.execute(
-            "UPDATE edges SET resolution_state = 3 WHERE id = ?1 AND resolution_state = 0",
-            params![edge_id],
-        )?;
-        Ok(())
-    }
-
-    /// Reset `resolution_state` from {2, 3} → 0 for edges whose target_name is in `names`.
-    ///
-    /// Called from the indexer when new symbols are added: an edge that was
-    /// previously "unresolvable" (no symbol with this name existed) or
-    /// "external" (target lived outside the index — but the user just vendored
-    /// it in-tree) may now be resolvable against the freshly-added target.
-    /// Returns the number of edges reopened. No-op when `names` is empty.
-    ///
-    /// tx-safe: single statement. Names are batched to honor SQLite's default
-    /// 999-parameter limit; only rows at state 2 or 3 are touched so the write
-    /// set stays tiny even on a large rename.
-    pub fn reset_unresolvable_for_names(&self, names: &[String]) -> Result<u32> {
-        if names.is_empty() {
-            return Ok(0);
-        }
-        const CHUNK: usize = 500;
-        let mut total: u32 = 0;
-        for chunk in names.chunks(CHUNK) {
-            let placeholders = vec!["?"; chunk.len()].join(",");
-            let sql = format!(
-                "UPDATE edges
-                 SET resolution_state = 0
-                 WHERE resolution_state IN (2, 3)
-                   AND target_name IN ({placeholders})"
-            );
-            let params: Vec<&dyn rusqlite::ToSql> =
-                chunk.iter().map(|n| n as &dyn rusqlite::ToSql).collect();
-            let n = self.conn.execute(&sql, params.as_slice())?;
-            total += n as u32;
-        }
-        Ok(total)
-    }
-}
+// The `Database` inherent impl is split across `store/` submodules for
+// navigability; each file holds one cohesive cluster of methods.
+mod store;
 
 /// An unresolved edge from the database (used by LSP resolution).
 #[derive(Debug, Clone)]
